+++ android/src/main/kotlin/com/example/verbuildno/NativeLib.kt
@@ -1,0 +1,10 @@
+package com.example.verbuildno
+
+object NativeLib {
+    init {
+        System.loadLibrary("native-lib")
+    }
+
+    external fun stringFromJNI(): String
+    external fun numberOfCoreFromJNI(): String
+}