+++ ios/Classes/Helper.h
@@ -1,0 +1,5 @@
+#import <Foundation/Foundation.h>
+
+@interface Helper : NSObject
+    -(NSNumber*) numberOfCore;
+@end