+++ ios/Classes/Helper.mm
@@ -1,0 +1,9 @@
+#import "Helper.h"
+#include "NativeLib.h"
+
+@implementation Helper
+-(NSNumber*) numberOfCore {
+    int core = NativeLib().numberOfCore();
+    return [NSNumber numberWithInt: core];
+}
+@end