+++ ios/Classes/NativeLib.cpp
@@ -1,0 +1,6 @@
+#include "NativeLib.h"
+#include <thread>
+
+int NativeLib::numberOfCore() {
+    return std::thread::hardware_concurrency();
+}