+++ ios/Classes/NativeLib.h
@@ -1,0 +1,12 @@
+#ifndef Test_hpp
+#define Test_hpp
+
+#include <stdio.h>
+#include <string>
+using namespace std;
+
+class NativeLib {
+    public:
+      int numberOfCore();
+};
+#endif