+++ lib/verbuildno.dart
@@ -1,4 +1,3 @@
-
 import 'verbuildno_platform_interface.dart';
 
 class Verbuildno {
@@ -8,5 +7,9 @@ class Verbuildno {
 
   Future<String?> getBundleIdentifer() {
     return VerbuildnoPlatform.instance.getBundleIdentifer();
+  }
+
+  Future<String?> getNumberOfCore() {
+    return VerbuildnoPlatform.instance.getNumberOfCore();
   }
 }